requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.6
requests>=2.31
numpy>=1.24
pillow>=10.0
